+++ svd-encoder/src/config.rs
@@ -4,7 +4,7 @@ use convert_case::{Boundary, Case, Casing};
 
 use crate::svd::BitRangeType;
 
-#[derive(Clone, Copy, Debug)]
+#[derive(Clone, Copy, Debug, PartialEq, Eq)]
 pub enum IdentifierFormat {
     /// `Camel` case
     ///
@@ -63,7 +63,7 @@ pub fn change_case(s: &str, case: Option<IdentifierFormat>) -> String {
     }
 }
 
-#[derive(Clone, Copy, Debug)]
+#[derive(Clone, Copy, Debug, PartialEq, Eq)]
 pub enum NumberFormat {
     /// `UpperHex` format
     ///
@@ -150,7 +150,7 @@ where
     }
 }
 
-#[derive(Clone, Copy, Debug)]
+#[derive(Clone, Copy, Debug, PartialEq, Eq)]
 pub struct FieldBitRangeFormat(pub BitRangeType);
 
 impl FromStr for FieldBitRangeFormat {
@@ -163,6 +163,93 @@ impl FromStr for FieldBitRangeFormat {
             "MsbLsb" => Ok(FieldBitRangeFormat(BitRangeType::MsbLsb)),
             _ => Err(()),
         }
+    }
+}
+
+#[derive(Clone, Copy, Debug, PartialEq, Eq)]
+#[non_exhaustive]
+/// Apply a kind of sorting
+pub enum Sorting {
+    /// Sort by addresses of offsets
+    Offset,
+    /// Same as [`Sorting::Offset`], but reversed
+    OffsetReversed,
+    /// Sort by name
+    Name,
+}
+
+impl Sorting {
+    fn from_parts(parts: &[&str]) -> Option<Self> {
+        if parts.contains(&"Offset") {
+            Some(Self::Offset)
+        } else if parts.contains(&"OffsetReserved") {
+            Some(Self::OffsetReversed)
+        } else if parts.contains(&"Name") {
+            Some(Self::Name)
+        } else {
+            None
+        }
+    }
+}
+
+#[derive(Clone, Copy, Debug, PartialEq, Eq)]
+pub enum DerivableSorting {
+    Unchanged(Option<Sorting>),
+    DeriveLast(Option<Sorting>),
+}
+
+impl DerivableSorting {
+    fn from_parts(parts: &[&str]) -> Self {
+        let sorting = Sorting::from_parts(parts);
+        if parts.contains(&"DerivedLast") {
+            Self::DeriveLast(sorting)
+        } else {
+            Self::Unchanged(sorting)
+        }
+    }
+}
+
+impl FromStr for DerivableSorting {
+    type Err = ();
+
+    fn from_str(s: &str) -> Result<Self, Self::Err> {
+        let parts = s.split(',').collect::<Vec<_>>();
+        Ok(DerivableSorting::from_parts(&parts))
+    }
+}
+
+impl Default for DerivableSorting {
+    fn default() -> Self {
+        Self::Unchanged(None)
+    }
+}
+
+#[derive(Clone, Copy, Debug, PartialEq, Eq)]
+pub enum RcSorting {
+    Unchanged(DerivableSorting),
+    RegistersFirst(DerivableSorting),
+    ClustersFirst(DerivableSorting),
+}
+
+impl Default for RcSorting {
+    fn default() -> Self {
+        Self::Unchanged(Default::default())
+    }
+}
+
+impl FromStr for RcSorting {
+    type Err = ();
+
+    fn from_str(s: &str) -> Result<Self, Self::Err> {
+        let parts = s.split(',').collect::<Vec<_>>();
+        let derivable_sorting = DerivableSorting::from_parts(&parts);
+        Ok(if parts.contains(&"RegistersFirst") {
+            Self::RegistersFirst(derivable_sorting)
+        } else if parts.contains(&"ClustersFirst") {
+            Self::ClustersFirst(derivable_sorting)
+        } else {
+            Self::Unchanged(derivable_sorting)
+        })
     }
 }
 
@@ -183,6 +270,9 @@ pub struct Config {
     ///
     /// format: hex, dec
     pub peripheral_base_address: NumberFormat,
+
+    /// Sort peripherals in specified order
+    pub peripheral_sorting: DerivableSorting,
 
     /// Format of addressBlock's offset element
     ///
@@ -208,6 +298,9 @@ pub struct Config {
     ///
     /// format: hex, dec
     pub cluster_address_offset: NumberFormat,
+
+    /// Sort registers and clusters in specified order
+    pub register_cluster_sorting: RcSorting,
 
     /// Format of register's name-kind elements
     /// - `derivedFrom`
@@ -245,6 +338,9 @@ pub struct Config {
     /// `None` means keep the original bitRange
     pub field_bit_range: Option<FieldBitRangeFormat>,
 
+    /// Sort fields in specified order
+    pub field_sorting: DerivableSorting,
+
     /// Format of enumeratedValues's name-kind elements
     /// - `derivedFrom`
     /// - `name`
@@ -278,11 +374,13 @@ impl Default for Config {
         Self {
             peripheral_name: None,
             peripheral_base_address: NumberFormat::UpperHex8,
+            peripheral_sorting: Default::default(),
             address_block_offset: NumberFormat::UpperHex,
             address_block_size: NumberFormat::UpperHex,
             interrupt_name: None,
             cluster_name: None,
             cluster_address_offset: NumberFormat::UpperHex,
+            register_cluster_sorting: Default::default(),
             register_name: None,
             register_address_offset: NumberFormat::UpperHex,
             register_size: NumberFormat::LowerHex,
@@ -290,6 +388,7 @@ impl Default for Config {
             register_reset_mask: NumberFormat::UpperHex16,
             field_name: None,
             field_bit_range: None,
+            field_sorting: Default::default(),
             enumerated_values_name: None,
             enumerated_value_name: None,
             enumerated_value_value: NumberFormat::Dec,
@@ -308,11 +407,13 @@ impl Config {
         match name {
             "peripheral_name" => self.peripheral_name = Some(value.parse().unwrap()),
             "peripheral_base_address" => self.peripheral_base_address = value.parse().unwrap(),
+            "peripheral_sorting" => self.peripheral_sorting = value.parse().unwrap(),
             "address_block_offset" => self.address_block_offset = value.parse().unwrap(),
             "address_block_size" => self.address_block_size = value.parse().unwrap(),
             "interrupt_name" => self.interrupt_name = Some(value.parse().unwrap()),
             "cluster_name" => self.cluster_name = Some(value.parse().unwrap()),
             "cluster_address_offset" => self.cluster_address_offset = value.parse().unwrap(),
+            "register_cluster_sorting" => self.register_cluster_sorting = value.parse().unwrap(),
             "register_name" => self.register_name = Some(value.parse().unwrap()),
             "register_address_offset" => self.register_address_offset = value.parse().unwrap(),
             "register_size" => self.register_size = value.parse().unwrap(),
@@ -320,6 +421,7 @@ impl Config {
             "register_reset_mask" => self.register_reset_mask = value.parse().unwrap(),
             "field_name" => self.field_name = Some(value.parse().unwrap()),
             "field_bit_range" => self.field_bit_range = Some(value.parse().unwrap()),
+            "field_sorting" => self.field_sorting = value.parse().unwrap(),
             "enumerated_values_name" => self.enumerated_values_name = Some(value.parse().unwrap()),
             "enumerated_value_name" => self.enumerated_value_name = Some(value.parse().unwrap()),
             "enumerated_value_value" => self.enumerated_value_value = value.parse().unwrap(),
@@ -345,6 +447,14 @@ impl Config {
     /// format: hex, dec
     pub fn peripheral_base_address(mut self, val: NumberFormat) -> Self {
         self.peripheral_base_address = val;
+        self
+    }
+
+    /// Sort peripherals in specified order
+    ///
+    /// `None` means keep the original order
+    pub fn peripheral_sorting(mut self, val: DerivableSorting) -> Self {
+        self.peripheral_sorting = val;
         self
     }
 
@@ -381,6 +491,14 @@ impl Config {
     /// format: hex, dec
     pub fn cluster_address_offset(mut self, val: NumberFormat) -> Self {
         self.cluster_address_offset = val;
+        self
+    }
+
+    /// Sort registers and clusters in specified order
+    ///
+    /// `None` means keep the original order
+    pub fn register_cluster_sorting(mut self, val: RcSorting) -> Self {
+        self.register_cluster_sorting = val;
         self
     }
 
@@ -433,6 +551,14 @@ impl Config {
     /// `None` means keep the original bitRange
     pub fn field_bit_range(mut self, val: Option<FieldBitRangeFormat>) -> Self {
         self.field_bit_range = val;
+        self
+    }
+
+    /// Sort fields in specified order
+    ///
+    /// `None` means keep the original order
+    pub fn field_sorting(mut self, val: DerivableSorting) -> Self {
+        self.field_sorting = val;
         self
     }