+++ svd-encoder/src/device.rs
@@ -1,5 +1,10 @@
+use svd_rs::Peripheral;
+
 use super::{new_node, Config, Element, Encode, EncodeChildren, EncodeError, XMLNode};
-use crate::svd::Device;
+use crate::{
+    config::{DerivableSorting, Sorting},
+    svd::Device,
+};
 
 impl Encode for Device {
     type Error = EncodeError;
@@ -57,11 +62,19 @@ impl Encode for Device {
                 .encode_with_config(config)?,
         );
 
-        let peripherals: Result<Vec<_>, _> = self
-            .peripherals
-            .iter()
-            .map(|peripheral| peripheral.encode_node_with_config(config))
-            .collect();
+        let peripherals: Result<Vec<_>, _> =
+            if config.peripheral_sorting == DerivableSorting::Unchanged(None) {
+                self.peripherals
+                    .iter()
+                    .map(|peripheral| peripheral.encode_node_with_config(config))
+                    .collect()
+            } else {
+                sort_derived_peripherals(&self.peripherals, config.peripheral_sorting)
+                    .into_iter()
+                    .map(|peripheral| peripheral.encode_node_with_config(config))
+                    .collect()
+            };
+
         elem.children.push({
             let mut e = Element::new("peripherals");
             e.children = peripherals?;
@@ -78,5 +91,44 @@ impl Encode for Device {
         );
 
         Ok(elem)
+    }
+}
+
+fn sort_peripherals(refs: &mut [&Peripheral], sorting: Option<Sorting>) {
+    if let Some(sorting) = sorting {
+        match sorting {
+            Sorting::Offset => refs.sort_by_key(|p| p.base_address),
+            Sorting::OffsetReversed => {
+                refs.sort_by_key(|p| -(p.base_address as i32));
+            }
+            Sorting::Name => refs.sort_by_key(|p| &p.name),
+        }
+    }
+}
+
+fn sort_derived_peripherals(
+    peripherals: &[Peripheral],
+    sorting: DerivableSorting,
+) -> Vec<&Peripheral> {
+    match sorting {
+        DerivableSorting::Unchanged(sorting) => {
+            let mut refs = peripherals.iter().collect::<Vec<_>>();
+            sort_peripherals(&mut refs, sorting);
+            refs
+        }
+        DerivableSorting::DeriveLast(sorting) => {
+            let mut common_refs = peripherals
+                .iter()
+                .filter(|p| p.derived_from.is_none())
+                .collect::<Vec<_>>();
+            let mut derived_refs = peripherals
+                .iter()
+                .filter(|p| p.derived_from.is_some())
+                .collect::<Vec<_>>();
+            sort_peripherals(&mut common_refs, sorting);
+            sort_peripherals(&mut derived_refs, sorting);
+            common_refs.extend(derived_refs);
+            common_refs
+        }
     }
 }