+++ svd-encoder/src/peripheral.rs
@@ -1,9 +1,11 @@
+use svd_rs::RegisterCluster;
+
 use super::{
     new_node, Config, Element, ElementMerge, Encode, EncodeChildren, EncodeError, XMLNode,
 };
 
 use crate::{
-    config::{change_case, format_number},
+    config::{change_case, format_number, DerivableSorting, RcSorting, Sorting},
     svd::{Peripheral, PeripheralInfo},
 };
 
@@ -102,10 +104,22 @@ impl Encode for PeripheralInfo {
         elem.children.append(&mut interrupts?);
 
         if let Some(v) = &self.registers {
-            let children: Result<Vec<_>, _> = v
-                .iter()
-                .map(|e| e.encode_node_with_config(config))
-                .collect();
+            let children: Result<Vec<_>, _> = match config.register_cluster_sorting {
+                RcSorting::Unchanged(DerivableSorting::Unchanged(None)) => v
+                    .iter()
+                    .map(|e| e.encode_node_with_config(config))
+                    .collect(),
+                RcSorting::Unchanged(sorting) => sort_derived_register_cluster(v, sorting)
+                    .into_iter()
+                    .map(|e| e.encode_node_with_config(config))
+                    .collect(),
+                RcSorting::RegistersFirst(sorting) => rc_sort(v, sorting, true)
+                    .map(|e| e.encode_node_with_config(config))
+                    .collect(),
+                RcSorting::ClustersFirst(sorting) => rc_sort(v, sorting, false)
+                    .map(|e| e.encode_node_with_config(config))
+                    .collect(),
+            };
 
             elem.children.push({
                 let mut e = Element::new("registers");
@@ -122,5 +136,68 @@ impl Encode for PeripheralInfo {
         }
 
         Ok(elem)
+    }
+}
+
+fn sort_register_cluster(refs: &mut [&RegisterCluster], sorting: Option<Sorting>) {
+    if let Some(sorting) = sorting {
+        match sorting {
+            Sorting::Offset => refs.sort_by_key(|r| r.address_offset()),
+            Sorting::OffsetReversed => {
+                refs.sort_by_key(|r| -(r.address_offset() as i32));
+            }
+            Sorting::Name => refs.sort_by_key(|r| r.name()),
+        }
+    }
+}
+
+fn sort_derived_register_cluster<'a>(
+    rcs: impl IntoIterator<Item = &'a RegisterCluster>,
+    sorting: DerivableSorting,
+) -> Vec<&'a RegisterCluster> {
+    match sorting {
+        DerivableSorting::Unchanged(sorting) => {
+            let mut refs = rcs.into_iter().collect::<Vec<_>>();
+            sort_register_cluster(&mut refs, sorting);
+            refs
+        }
+        DerivableSorting::DeriveLast(sorting) => {
+            let mut common_refs = Vec::new();
+            let mut derived_refs = Vec::new();
+            for rc in rcs {
+                if rc.derived_from().is_some() {
+                    derived_refs.push(rc);
+                } else {
+                    common_refs.push(rc);
+                }
+            }
+            sort_register_cluster(&mut common_refs, sorting);
+            sort_register_cluster(&mut derived_refs, sorting);
+            common_refs.extend(derived_refs);
+            common_refs
+        }
+    }
+}
+
+fn rc_sort(
+    v: &[RegisterCluster],
+    sorting: DerivableSorting,
+    register_first: bool,
+) -> impl Iterator<Item = &RegisterCluster> {
+    let reg_refs = v
+        .iter()
+        .filter(|rc| matches!(rc, RegisterCluster::Register(_)))
+        .collect::<Vec<_>>();
+    let reg_refs = sort_derived_register_cluster(reg_refs, sorting);
+
+    let c_refs = v
+        .iter()
+        .filter(|rc| matches!(rc, RegisterCluster::Cluster(_)))
+        .collect::<Vec<_>>();
+    let c_refs = sort_derived_register_cluster(c_refs, sorting);
+    if register_first {
+        reg_refs.into_iter().chain(c_refs.into_iter())
+    } else {
+        c_refs.into_iter().chain(reg_refs.into_iter())
     }
 }