+++ svd-encoder/src/register.rs
@@ -1,9 +1,11 @@
+use svd_rs::Field;
+
 use super::{
     new_node, Config, Element, ElementMerge, Encode, EncodeChildren, EncodeError, XMLNode,
 };
 
 use crate::{
-    config::{change_case, format_number},
+    config::{change_case, format_number, DerivableSorting, Sorting},
     svd::{Register, RegisterInfo},
 };
 
@@ -74,10 +76,19 @@ impl Encode for RegisterInfo {
         }
 
         if let Some(v) = &self.fields {
-            let children = v
-                .iter()
-                .map(|field| field.encode_node_with_config(config))
-                .collect::<Result<Vec<_>, EncodeError>>()?;
+            let children: Result<Vec<_>, _> =
+                if config.field_sorting == DerivableSorting::Unchanged(None) {
+                    v.iter()
+                        .map(|field| field.encode_node_with_config(config))
+                        .collect()
+                } else {
+                    sort_derived_fields(v, config.field_sorting)
+                        .into_iter()
+                        .map(|field| field.encode_node_with_config(config))
+                        .collect()
+                };
+
+            let children = children?;
             if !children.is_empty() {
                 let mut fields = Element::new("fields");
                 fields.children = children;
@@ -93,5 +104,53 @@ impl Encode for RegisterInfo {
         }
 
         Ok(elem)
+    }
+}
+
+fn sort_fields(refs: &mut [&Field], sorting: Option<Sorting>) {
+    if let Some(sorting) = sorting {
+        match sorting {
+            Sorting::Offset => refs.sort_by_key(|f| f.bit_offset()),
+            Sorting::OffsetReversed => {
+                refs.sort_by_key(|f| -(f.bit_offset() as i32));
+            }
+            Sorting::Name => refs.sort_by_key(|f| &f.name),
+        }
+    }
+}
+
+fn sort_derived_fields(v: &[Field], sorting: DerivableSorting) -> Vec<&Field> {
+    match sorting {
+        DerivableSorting::Unchanged(sorting) => {
+            let mut refs = v.iter().collect::<Vec<_>>();
+            sort_fields(&mut refs, sorting);
+            refs
+        }
+        DerivableSorting::DeriveLast(sorting) => {
+            let mut common_refs = Vec::with_capacity(v.len());
+            let mut derived_refs = Vec::new();
+            for f in v.iter() {
+                if f.derived_from.is_some() {
+                    derived_refs.push(f);
+                } else {
+                    let mut derived = false;
+                    for ev in &f.enumerated_values {
+                        if ev.derived_from.is_some() {
+                            derived_refs.push(f);
+                            derived = true;
+                            break;
+                        }
+                    }
+                    if !derived {
+                        common_refs.push(f);
+                    }
+                }
+            }
+            sort_fields(&mut common_refs, sorting);
+            sort_fields(&mut derived_refs, sorting);
+            common_refs.extend(derived_refs);
+
+            common_refs
+        }
     }
 }